Fix bugs and refactor
