+++ Source/DafnyCore/AST/Expressions/Variables/IVariableContracts.cs
@@ -81,7 +81,7 @@ public abstract class IVariableContracts : TokenNode, IVariable {
 
   public abstract IToken NameToken { get; }
   public DafnySymbolKind Kind => throw new NotImplementedException();
-  public string GetHoverText(DafnyOptions options) {
+  public string GetDescription(DafnyOptions options) {
     throw new NotImplementedException();
   }
 }
